Change Id data type to long

+++ hw-006/src/main/java/ru/glavs/hw005/dao/AuthorDao.java
@@ -5,7 +5,7 @@ import ru.glavs.hw005.domain.Author;
 import java.util.List;
 
 public interface AuthorDao {
-    Author getById(int id);
+    Author getById(long id);
 
     List<Author> getAll();
